New article in a possible new section
Added a new sections called "Marketing and Advertisement". Not sure about the title, but it covers a few things and not just marketing, you know? The article I added is about how to grow your network and spread the word of your event without being a big company running events, but an individual and using and working with your attendees and non monetary business models.